added InterleavedRBFitter class to fitter.py
