+++ qiskit/ignis/verification/randomized_benchmarking/__init__.py
@@ -21,6 +21,5 @@ Measurement correction module
 from .Clifford import Clifford
 from . import clifford_utils
 from .circuits import randomized_benchmarking_seq
-from .fitters import RBFitter
-from .interleaved_fitters import InterleavedRBFitter
+from .fitters import RBFitter, InterleavedRBFitter
 from . import rb_utils
